LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/windows/noreferrer-null-opener-expected.txt

CONSOLE MESSAGE: QuotaExceededError: The quota has been exceeded.

Harness Error (TIMEOUT), message = null

TIMEOUT rel=noreferrer nullifies window.opener Test timed out
